Rename `ImageDownloadable` to `ImageDownloader` and refactor mock image downloaders.
