+++ WooCommerce/WooCommerceTests/Mockups/MockImageDownloader.swift
@@ -1,6 +1,6 @@
 @testable import Kingfisher
 
-final class MockImageDownloader: ImageDownloader {
+final class MockKingfisherImageDownloader: Kingfisher.ImageDownloader {
     // Mocks in-memory cache.
     private let imagesByKey: [String: UIImage]
